+++ packages/agenda/lib/src/agenda/components/navigation/navigation.component.ts
@@ -8,7 +8,7 @@ import {
 	OnInit,
 	OnDestroy,
 } from '@angular/core';
-import { Subject } from 'rxjs/Subject';
+import { Subject } from 'rxjs';
 import { takeUntil, debounceTime, distinctUntilChanged } from 'rxjs/operators';
 
 import { VIEWS } from '../../types/agenda.types';