+++ packages/context/lib/src/context/store/context/context.actioncreator.ts
@@ -1,7 +1,6 @@
 import { Injectable } from '@angular/core';
 import { NgRedux } from '@angular-redux/store';
-import { Subscription } from 'rxjs';
-import { Observable } from 'rxjs';
+import { Subscription, Observable } from 'rxjs';
 
 import { Context } from '../../types/context.types';
 import { ContextState } from '../store.types';