+++ packages/flyout/lib/src/flyout/directives/flyout.directive.spec.ts
@@ -7,7 +7,7 @@ import { FlyoutZoneDirective } from './flyout-zone.directive';
 
 import { FlyoutService } from '../services/flyout.service';
 
-import { Subject } from 'rxjs/Subject';
+import { Subject } from 'rxjs';
 
 class MockFlyoutService {
 	// Observable string sources
@@ -98,7 +98,6 @@ describe('Flyout directive without flyout zone', () => {
 		componentDebugElement = fixture.debugElement.query(By.directive(FlyoutDirective));
 		componentElement = <HTMLElement>componentDebugElement.nativeElement;
 	}));
-
 
 	it('should open and close', () => {
 		comp.element.open();