+++ packages/logo/lib/src/logo/logo.module.ts
@@ -3,7 +3,6 @@ import { CommonModule } from '@angular/common';
 
 import { Components } from './components/index';
 
-
 @NgModule({
 	imports: [
 		CommonModule,